backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
pydantic[email]>=2.4.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-groq>=0.1.1
groq>=0.4.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.12.0
python-multipart>=0.0.6
tenacity>=8.2.3
loguru>=0.7.2